[FEAT] 지도 관련 기능 추가

See merge request s09-webmobile4-sub2/S09P12D104!11

+++ frontend/android/app/src/main/java/com/ssafy/hifes/data/model/GroupDto.kt
@@ -9,7 +9,7 @@ data class GroupDto(
     var createdAt: Date,
     var maxPop: Int,
     var content: String,
-    var getterLatitude: Float,
-    var getterLongitude: Float,
+    var getterLatitude: Double,
+    var getterLongitude: Double,
     var groupPassword: Int,
 )
